more flexible request filter handling
Filters are now passed as array and can contain the whole filter
syntax (but are limited to AND filters currently).

Filters can now be passed via POST and GET method

This partly reverts the patch by nikopp as tagclouds again replace
all request filters (but not the filters set in a table itself).