Restructure page and title types
The types page and pageid are now handled similar, but pageid is not
cleaned. If page and pageid are provided without title then the id is
shown as title. If they are provided with a title, then this title is
shown.
If a title field is not provided with a title, then the shown title depends on
the useheadings setting. If it has a title, then the title is shown.